GuitarNeckComponent: Move from _common to midiChords
It was only used there anyway. Furthermore, this move allows for the
juce_audio_basics dependency to be moved from _common into midiChords
and in result, save compilation time in the other plugins.
