+++ pizjuce/midiChords/CMakeLists.txt
@@ -10,12 +10,14 @@ piz_add_plugin(
 target_sources(
   piz_midiChords
   PRIVATE # Headers
+          "${CMAKE_CURRENT_LIST_DIR}/GuitarNeckComponent.h"
           "${CMAKE_CURRENT_LIST_DIR}/MidiChords.h"
           "${CMAKE_CURRENT_LIST_DIR}/MidiChordsEditor.h"
           # Sources
+          "${CMAKE_CURRENT_LIST_DIR}/GuitarNeckComponent.cpp"
           "${CMAKE_CURRENT_LIST_DIR}/MidiChords.cpp"
           "${CMAKE_CURRENT_LIST_DIR}/MidiChordsEditor.cpp")
 
 target_link_libraries(
-  piz_midiChords PRIVATE juce::juce_audio_utils juce::juce_cryptography
-                         juce::juce_gui_basics)
+  piz_midiChords PRIVATE juce::juce_audio_basics juce::juce_audio_utils
+                         juce::juce_cryptography juce::juce_gui_basics)
